# Optimize Qwen3-moe model by using flashinfer fused allreduce 