<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

<!-- Describe the purpose and goals of this pull request. -->
This PR is to make qwen moe model leverage FlashInfer fused_allreduce to fuse allreduce+rmsnorm+residual_add. 
The E2E performance improved. Currently use a small MoE model so the speedup is slightly, larger 235B model can gain better performance. Will try and update the report.


## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

The result is correct. No accuracy drop.
```
$python test_openai.py 
ChatCompletion(id='ab6d5c2a3da74d58908895b51a64ebd2', choices=[Choice(finish_reason='length', index=0, logprobs=None, message=ChatCompletionMessage(content="<think>\nOkay, the user asked for three countries and their capitals, and then how I rank them. Let me start by picking three countries. Maybe the US, Japan, and Brazil. Their capitals are Washington, D.C., Tokyo, and Brasília. Now, how to rank them? The user didn't specify the criteria, so I need to think of possible ways. Maybe by population, economic size, or cultural influence. Let me check the population. The US has around 330 million, Japan about 125 million, Brazil 215 million. So US first, Brazil second, Japan third. But if I consider GDP, the US is the largest, then Japan, then Brazil. Alternatively, cultural influence: Japan has a strong cultural impact, maybe higher than Brazil. But the user might not have a specific criteria. I should mention that the ranking depends on the criteria. Maybe list different rankings based on different factors. That way, the answer is comprehensive", refusal=None, role='assistant', annotations=None, audio=None, function_call=None, tool_calls=None, reasoning_content=None), matched_stop=None)], created=1756901402, model='default', object='chat.completion', service_tier=None, system_fingerprint=None, usage=CompletionUsage(completion_tokens=200, prompt_tokens=33, total_tokens=233, completion_tokens_details=None, prompt_tokens_details=None, reasoning_tokens=0), metadata={'weight_version': 'default'})
```

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

## Benchmarking and Profiling

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->

```
==========This PR==========
Server run:
python3 -m sglang.launch_server --model-path /home/admin/Qwen3-30B-A3B --attention-backend flashinfer --trust-remote-code --tp-size=8 --disable-radix-cache --enable-flashinfer-allreduce-fusion

Benchmark command:
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 20 --max-concurrency 1 --output-file opt_ar.jsonl --port 30000 --warmup-requests 10
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 200 --max-concurrency 32 --output-file opt_ar.jsonl --port 30000 --warmup-requests 10
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 300 --max-concurrency 64 --output-file opt_ar.jsonl --port 30000 --warmup-requests 10
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 400 --max-concurrency 128 --output-file opt_ar.jsonl --port 30000 --warmup-requests 10

$python test/srt/parse_results.py opt_ar.jsonl 

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |          51635.460 |              12.909 |         76.926 |           75.256 |        89.571 |          0.000 |            0.000 |         0.000 |                12.909 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |            32.000 |          66471.176 |              16.618 |       1786.338 |         1932.208 |      2035.871 |          0.000 |            0.000 |         0.000 |                 0.519 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            64.000 |          66759.245 |              16.690 |       3452.846 |         3799.681 |      3949.549 |          0.000 |            0.000 |         0.000 |                 0.261 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |           128.000 |          65622.899 |              16.406 |       6602.345 |         7596.307 |      7899.919 |          0.000 |            0.000 |         0.000 |                 0.128 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+


Main:

python3 -m sglang.launch_server --model-path /home/admin/Qwen3-30B-A3B --attention-backend flashinfer --trust-remote-code --tp-size=8 --disable-radix-cache

curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 20 --max-concurrency 1 --output-file main.jsonl --port 30000 --warmup-requests 10
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 200 --max-concurrency 32 --output-file main.jsonl --port 30000 --warmup-requests 10
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 300 --max-concurrency 64 --output-file main.jsonl --port 30000 --warmup-requests 10
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 4000 --random-output-len 1 --random-range-ratio 1 --num-prompts 400 --max-concurrency 128 --output-file main.jsonl --port 30000 --warmup-requests 10

$python test/srt/parse_results.py main.jsonl 

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |          51063.302 |              12.766 |         77.780 |           75.948 |        90.029 |          0.000 |            0.000 |         0.000 |                12.766 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |            32.000 |          66213.827 |              16.553 |       1794.221 |         1941.204 |      2012.882 |          0.000 |            0.000 |         0.000 |                 0.517 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            64.000 |          66627.148 |              16.657 |       3460.273 |         3808.979 |      3949.841 |          0.000 |            0.000 |         0.000 |                 0.260 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |           128.000 |          65931.725 |              16.483 |       6555.403 |         7596.530 |      7773.423 |          0.000 |            0.000 |         0.000 |                 0.129 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+

```

## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).