@BBuf 2025-09-03
Can you add a gsm8k acc test and torch profiler result?