@yuan-luo 2025-09-04
```
PR:
$python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000 --data-path ./test.jsonl
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:10<00:00, 18.23it/s]
Accuracy: 0.730
Invalid: 0.000
Latency: 11.144 s
Output throughput: 4556.411 token/s

Main:
$python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000 --data-path ./test.jsonl
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:11<00:00, 16.71it/s]
Accuracy: 0.695
Invalid: 0.000
Latency: 12.098 s
Output throughput: 4544.950 token/s
```